+++ CHANGES.md
@@ -119,6 +119,8 @@
     -   Do not fail displaying inline images on empty links. ([GH-320][])
     -   Fix off-by-one error in `markdown-inline-code-at-pos`.
         ([GH-313][])
+    -   Fix bounds during inline comment syntax propertization. ([GH-327][])
+
   [gh-349]: https://github.com/jrblevin/markdown-mode/issues/349]
   [gh-171]: https://github.com/jrblevin/markdown-mode/issues/171
   [gh-216]: https://github.com/jrblevin/markdown-mode/issues/216
@@ -163,6 +165,7 @@
   [gh-317]: https://github.com/jrblevin/markdown-mode/pull/317
   [gh-320]: https://github.com/jrblevin/markdown-mode/pull/320
   [gh-322]: https://github.com/jrblevin/markdown-mode/pull/322
+  [gh-327]: https://github.com/jrblevin/markdown-mode/issues/327
   [gh-331]: https://github.com/jrblevin/markdown-mode/issues/331
   [gh-335]: https://github.com/jrblevin/markdown-mode/pull/335
   [gh-340]: https://github.com/jrblevin/markdown-mode/issues/340